25/10

+++ src/bomberman/graphics/Sprite.java
@@ -181,7 +181,7 @@ public class Sprite {
         return x2;
     }
 
-    public static Sprite movingSprite(Sprite x1, Sprite x2, int animate, int time) {
+    public static Sprite movingSprite(Sprite x1, Sprite x2, long animate, int time) {
         int diff = time / 2;
         return (animate % time > diff) ? x1 : x2;
     }
